+++ api/public/controllers.autoload/Transactions.php
@@ -220,7 +220,8 @@ class Transactions
             }
 
             /* Execute Operations */
-            $db = new EnsoDB(true);
+            $transactional = true;
+            $db = new EnsoDB($transactional);
             $db->getDB()->beginTransaction();
 
 
@@ -234,29 +235,29 @@ class Transactions
                 "accounts_account_to_id" => $accountTo,
                 "categories_category_id" => $categoryID,
                 "is_essential" => $isEssential,
-            ], true);
+            ], $transactional);
 
-            $userID = UserModel::getUserIdByName($authusername, true);
-            UserModel::setLastUpdateTimestamp($userID, time(), true);
+            $userID = UserModel::getUserIdByName($authusername, $transactional);
+            UserModel::setLastUpdateTimestamp($userID, time(), $transactional);
 
             switch ($type) {
                 case DEFAULT_TYPE_INCOME_TAG:
-                    AccountModel::setNewAccountBalance($accountTo,
-                        AccountModel::recalculateBalanceForAccountIncrementally($accountTo, $date_timestamp - 1, time() + 1, true),
-                        true);
+                    AccountModel::setNewAccountBalance($userID, $accountTo,
+                        AccountModel::recalculateBalanceForAccountIncrementally($accountTo, $date_timestamp - 1, time() + 1, $transactional),
+                        $transactional);
                     break;
                 case DEFAULT_TYPE_EXPENSE_TAG:
-                    AccountModel::setNewAccountBalance($accountFrom,
-                        AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, $date_timestamp - 1, time() + 1, true),
-                        true);
+                    AccountModel::setNewAccountBalance($userID, $accountFrom,
+                        AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, $date_timestamp - 1, time() + 1, $transactional),
+                        $transactional);
                     break;
                 case DEFAULT_TYPE_TRANSFER_TAG:
-                    AccountModel::setNewAccountBalance($accountTo,
-                        AccountModel::recalculateBalanceForAccountIncrementally($accountTo, $date_timestamp - 1, time() + 1, true),
-                        true);
-                    AccountModel::setNewAccountBalance($accountFrom,
-                        AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, $date_timestamp - 1, time() + 1, true),
-                        true);
+                    AccountModel::setNewAccountBalance($userID, $accountTo,
+                        AccountModel::recalculateBalanceForAccountIncrementally($accountTo, $date_timestamp - 1, time() + 1, $transactional),
+                        $transactional);
+                    AccountModel::setNewAccountBalance($userID, $accountFrom,
+                        AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, $date_timestamp - 1, time() + 1, $transactional),
+                        $transactional);
                     break;
             }
 
@@ -306,6 +307,7 @@ class Transactions
                 ["amount", "date_timestamp", "type", "accounts_account_from_id", "accounts_account_to_id"]
             )[0];
 
+
             $oldAmount = $trxObj["amount"];
             $oldTimestamp = $trxObj["date_timestamp"];
             $oldType = $trxObj["type"];
@@ -316,6 +318,14 @@ class Transactions
             if (isset($trxObj["accounts_account_from_id"]))
                 $oldAccountFrom = $trxObj["accounts_account_from_id"];
 
+            // Make sure accounts belong to user
+            if (!AccountModel::exists([
+                "users_user_id" => $userID,
+                "account_id" => $oldAccountFrom ?? $oldAccountTo,
+            ])) {
+                throw new BadValidationTypeException("Account not found!");
+            }
+
             TransactionModel::delete([
                 "transaction_id" => $trxID
             ], true);
@@ -324,22 +334,22 @@ class Transactions
             switch ($oldType) {
                 case DEFAULT_TYPE_INCOME_TAG:
                     // Decrement $oldAmount to level it out
-                    AccountModel::setNewAccountBalance($oldAccountTo,
+                    AccountModel::setNewAccountBalance($userID, $oldAccountTo,
                         AccountModel::recalculateBalanceForAccountIncrementally($oldAccountTo, $oldTimestamp - 1, time() + 1, true),
                         true);
 
                     break;
                 case DEFAULT_TYPE_EXPENSE_TAG:
                     // Increment $oldAmount to level it out, by reimbursing the amount
-                    AccountModel::setNewAccountBalance($oldAccountFrom,
+                    AccountModel::setNewAccountBalance($userID, $oldAccountFrom,
                         AccountModel::recalculateBalanceForAccountIncrementally($oldAccountFrom, $oldTimestamp - 1, time() + 1, true),
                         true);
                     break;
                 case DEFAULT_TYPE_TRANSFER_TAG:
-                    AccountModel::setNewAccountBalance($oldAccountFrom,
+                    AccountModel::setNewAccountBalance($userID, $oldAccountFrom,
                         AccountModel::recalculateBalanceForAccountIncrementally($oldAccountFrom, $oldTimestamp - 1, time() + 1, true),
                         true);
-                    AccountModel::setNewAccountBalance($oldAccountTo,
+                    AccountModel::setNewAccountBalance($userID, $oldAccountTo,
                         AccountModel::recalculateBalanceForAccountIncrementally($oldAccountTo, $oldTimestamp - 1, time() + 1, true),
                         true);
                     break;
@@ -494,6 +504,14 @@ class Transactions
             if (isset($trxObj["accounts_account_from_id"]))
                 $oldAccountFrom = $trxObj["accounts_account_from_id"];
 
+            // Make sure accounts belong to user
+            if (!AccountModel::exists([
+                "users_user_id" => $userID,
+                "account_id" => $oldAccountFrom ?? $oldAccountTo,
+            ])) {
+                throw new BadValidationTypeException("Account not found!");
+            }
+
             TransactionModel::editWhere([
                 "transaction_id" => $trxID,
             ], [
@@ -534,7 +552,7 @@ class Transactions
                 case DEFAULT_TYPE_INCOME_TAG:
                     // Decrement $oldAmount to level it out
                     //AccountModel::changeBalance($userID, $accountTo, $amount, true);
-                    AccountModel::setNewAccountBalance($accountTo,
+                    AccountModel::setNewAccountBalance($userID, $accountTo,
                         AccountModel::recalculateBalanceForAccountIncrementally($accountTo, min($oldTimestamp, $date_timestamp) - 1, time() + 1, true),
                         true);
 
@@ -542,17 +560,17 @@ class Transactions
                 case DEFAULT_TYPE_EXPENSE_TAG:
                     // Increment $oldAmount to level it out, by reimbursing the amount
                     //AccountModel::changeBalance($userID, $accountFrom, -$amount, true);
-                    AccountModel::setNewAccountBalance($accountFrom,
+                    AccountModel::setNewAccountBalance($userID, $accountFrom,
                         AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, min($oldTimestamp, $date_timestamp) - 1, time() + 1, true),
                         true);
                     break;
                 case DEFAULT_TYPE_TRANSFER_TAG:
                     //AccountModel::changeBalance($userID, $accountFrom, -$amount, true);
                     //AccountModel::changeBalance($userID, $accountTo, +$amount, true);
-                    AccountModel::setNewAccountBalance($accountTo,
+                    AccountModel::setNewAccountBalance($userID, $accountTo,
                         AccountModel::recalculateBalanceForAccountIncrementally($accountTo, min($oldTimestamp, $date_timestamp) - 1, time() + 1, true),
                         true);
-                    AccountModel::setNewAccountBalance($accountFrom,
+                    AccountModel::setNewAccountBalance($userID, $accountFrom,
                         AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, min($oldTimestamp, $date_timestamp) - 1, time() + 1, true),
                         true);
                     break;
@@ -575,21 +593,21 @@ class Transactions
             switch ($split_type) {
                 case DEFAULT_TYPE_INCOME_TAG:
                     // Decrement $oldAmount to level it out
-                    AccountModel::setNewAccountBalance($split_accountToID,
+                    AccountModel::setNewAccountBalance($userID, $split_accountToID,
                         AccountModel::recalculateBalanceForAccountIncrementally($split_accountToID, min($oldTimestamp, $date_timestamp) - 1, time() + 1, true), true);
                     break;
                 case DEFAULT_TYPE_EXPENSE_TAG:
                     // Increment $oldAmount to level it out, by reimbursing the amount
                     /*AccountModel::changeBalance($userID, $accountFrom, -$split_amount, true);*/
-                    AccountModel::setNewAccountBalance($split_accountFromID,
+                    AccountModel::setNewAccountBalance($userID, $split_accountFromID,
                         AccountModel::recalculateBalanceForAccountIncrementally($split_accountFromID, min($oldTimestamp, $date_timestamp) - 1, time() + 1,
                             true), true);
                     break;
                 case DEFAULT_TYPE_TRANSFER_TAG:
-                    AccountModel::setNewAccountBalance($split_accountToID,
+                    AccountModel::setNewAccountBalance($userID, $split_accountToID,
                         AccountModel::recalculateBalanceForAccountIncrementally($split_accountToID, min($oldTimestamp, $date_timestamp) - 1, time() + 1,
                             true), true);
-                    AccountModel::setNewAccountBalance($split_accountFromID,
+                    AccountModel::setNewAccountBalance($userID, $split_accountFromID,
                         AccountModel::recalculateBalanceForAccountIncrementally($split_accountFromID, min($oldTimestamp, $date_timestamp) - 1, time() + 1,
                             true), true);
                     break;
@@ -790,20 +808,20 @@ class Transactions
 
                 switch ($type) {
                     case DEFAULT_TYPE_INCOME_TAG:
-                        AccountModel::setNewAccountBalance($accountTo,
+                        AccountModel::setNewAccountBalance($userID, $accountTo,
                             AccountModel::recalculateBalanceForAccountIncrementally($accountTo, $date_timestamp - 1, time() + 1, true),
                             true);
                         break;
                     case DEFAULT_TYPE_EXPENSE_TAG:
-                        AccountModel::setNewAccountBalance($accountFrom,
+                        AccountModel::setNewAccountBalance($userID, $accountFrom,
                             AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, $date_timestamp - 1, time() + 1, true),
                             true);
                         break;
                     case DEFAULT_TYPE_TRANSFER_TAG:
-                        AccountModel::setNewAccountBalance($accountTo,
+                        AccountModel::setNewAccountBalance($userID, $accountTo,
                             AccountModel::recalculateBalanceForAccountIncrementally($accountTo, $date_timestamp - 1, time() + 1, true),
                             true);
-                        AccountModel::setNewAccountBalance($accountFrom,
+                        AccountModel::setNewAccountBalance($userID, $accountFrom,
                             AccountModel::recalculateBalanceForAccountIncrementally($accountFrom, $date_timestamp - 1, time() + 1, true),
                             true);
                         break;
